+++ src/frontend/components/helpers/output.ts
@@ -14,8 +14,7 @@ import { isMainWindow, isOutputWindow, newToast } from "../../utils/common"
 import { translateText } from "../../utils/language"
 import { confirmCustom } from "../../utils/popup"
 import { send } from "../../utils/request"
-import { sendBackgroundToStage } from "../../utils/stageTalk"
-import { hasStageStreamViewers } from "../../utils/stageStreamViewers"
+import { hasStageStreamViewers, sendBackgroundToStage } from "../../utils/stageTalk"
 import { TemplateHelper } from "../../utils/templates"
 import { videoExtensions } from "../../values/extensions"
 import { customActionActivation, runAction } from "../actions/actions"
