+++ src/server/stage/components/Stagebox.svelte
@@ -29,16 +29,13 @@
     $: currentSlide = currentOutput?.out?.slide || (slideOffset !== 0 ? $outputSlideCache[currentOutput?.id || ""] || null : null)
 
     // pick the pushed frame for this layout's output, falling back deterministically
-    // (legacy "alpha"/"default" keys come from the polled REQUEST_STREAM path)
-    function getStreamCapture(streamMap: any, outputId: string | undefined, alpha: boolean) {
+    function getStreamCapture(streamMap: any, outputId: string | undefined) {
         if (outputId && streamMap[outputId]) return streamMap[outputId]
 
-        const outputKeys = Object.keys(streamMap)
-            .filter((key) => key !== "alpha" && key !== "default")
-            .sort()
+        const outputKeys = Object.keys(streamMap).sort()
         if (outputKeys.length) return streamMap[outputKeys[0]]
 
-        return streamMap[alpha ? "alpha" : "default"]
+        return null
     }
 
     $: currentBackground = $background
@@ -200,7 +197,7 @@
         <span style="pointer-events: none;width: 100%;height: 100%;">
             {#if item.type === "current_output" || id.includes("current_output")}
                 <!-- width gets squished when resized -->
-                <PreviewCanvas outputId={stageLayout?.settings?.output} capture={getStreamCapture($stream, stageLayout?.settings?.output, id.includes("_alpha"))} />
+                <PreviewCanvas outputId={stageLayout?.settings?.output} capture={getStreamCapture($stream, stageLayout?.settings?.output)} />
             {:else if item.type === "slide_text" || id.includes("slide")}
                 {@const slideBackground = slideOffset === 0 ? currentBackground : slideOffset === 1 ? currentBackground.next : null}
